client-examples/example/py/requirements.txt
menlosystemcore
pywebchannel
websocket-client
websockets
